+++ contracts/modules/MerkleDropMinter.sol
@@ -34,7 +34,7 @@ contract MerkleDropMinter is IMerkleDropMinter, BaseMinter {
     function createEditionMint(
         address edition,
         bytes32 merkleRootHash,
-        uint256 price_,
+        uint128 price_,
         uint32 startTime,
         uint32 endTime,
         uint32 maxMintable_,
@@ -107,7 +107,7 @@ contract MerkleDropMinter is IMerkleDropMinter, BaseMinter {
     }
 
     /**
-     * @dev Returns the `EditionMintData` for `edition.
+     * @dev Returns the `EditionMintData` for `edition`.
      * @param edition Address of the song edition contract we are minting for.
      */
     function editionMintData(address edition, uint256 mintId) public view returns (EditionMintData memory) {
